+++ src/main/java/tornadofx/Component.kt
@@ -26,8 +26,7 @@ import java.util.logging.Logger
 import java.util.prefs.Preferences
 import kotlin.concurrent.thread
 import kotlin.properties.ReadOnlyProperty
-import kotlin.reflect.KClass
-import kotlin.reflect.KProperty
+import kotlin.reflect.*
 
 interface Injectable
 
@@ -123,11 +122,28 @@ abstract class Component {
     fun <T> background(func: () -> T) = task(func)
 
     /**
-     * Perform the given operation on an Injectable of the specified type, for example:
+     * Perform the given operation on an Injectable of the specified type asynchronousyly.
      *
      * MyController::class.runAsync { functionOnMyController() } ui { processResult(it) }
      */
     inline fun <reified T : Injectable, R> KClass<T>.runAsync(noinline op: T.() -> R) = task { op(find(T::class)) }
+
+    /**
+     * Perform the given operation on an Injectable class function member asynchronousyly.
+     *
+     * CustomerController::listContacts.runAsync(customerId).ui { processResult(it) }
+     */
+    inline fun <reified InjectableType : Injectable, reified ReturnType> KFunction1<InjectableType, ReturnType>.runAsync() = task { invoke(find(InjectableType::class)) }
+    /**
+     * Perform the given operation on an Injectable class function member asynchronousyly.
+     *
+     * CustomerController::listCustomers.runAsync().ui { processResult(it) }
+     */
+    inline fun <reified InjectableType : Injectable, reified P1, reified ReturnType> KFunction2<InjectableType, P1, ReturnType>.runAsync(p1: P1) = task { invoke(find(InjectableType::class), p1) }
+    inline fun <reified InjectableType : Injectable, reified P1, reified P2, reified ReturnType> KFunction3<InjectableType, P1, P2, ReturnType>.runAsync(p1: P1, p2: P2) = task { invoke(find(InjectableType::class), p1, p2) }
+    inline fun <reified InjectableType : Injectable, reified P1, reified P2, reified P3, reified ReturnType> KFunction4<InjectableType, P1, P2, P3, ReturnType>.runAsync(p1: P1, p2: P2, p3: P3) = task { invoke(find(InjectableType::class), p1, p2, p3) }
+    inline fun <reified InjectableType : Injectable, reified P1, reified P2, reified P3, reified P4, reified ReturnType> KFunction5<InjectableType, P1, P2, P3, P4, ReturnType>.runAsync(p1: P1, p2: P2, p3: P3, p4: P4) = task { invoke(find(InjectableType::class), p1, p2, p3, p4) }
+
     fun <T> runAsync(func: () -> T) = task(func)
     infix fun <T> Task<T>.ui(func: (T) -> Unit) = success(func)
 }