+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- `runAsync` on Injectable class references (`CustomerController::class.runAsync { listContacts(customerId) }`)
+- `runAsync` on Injectable class function references (`CustomerController::listContacts.runAsync(customerId)`)
 - `ObservableValue.onChange` listener
 - LayoutDebugger (https://github.com/edvin/tornadofx/wiki/Layout-Debugger)
 - ViewModel (https://github.com/edvin/tornadofx/wiki/ViewModel)